% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.pie.R
\name{portfolio.pie}
\alias{portfolio.pie}
\title{Plot a portfolio pie.}
\usage{
portfolio.pie(portfolio, scenario.set, main = "Portfolio")
}
\description{
\code{portfolio.pie} plots a pie chart of a portfolio given a specific scenario set
}

