% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webinar.cpom.R
\docType{package}
\name{webinar.cpom-package}
\alias{webinar.cpom}
\alias{webinar.cpom-package}
\title{Webinar - Contemporary Portfolio Optimization Modeling using R}
\description{
http://finance-r.com/cpom/
}
\details{
In the first part of this Webinar we review the most common ways to conduct 
the task of portfolio optimization with R. After this introduction, which can 
be of use for anyone interested in portfolio opimization in general, some 
remarks on the modeling of portfolio problems will be addressed. In the second 
part, a revolutionary way to model and solve portfolio optimization problems 
using R will be shown. The basic idea of conceptualizing a new way to model 
portfolio optimization problems is to build a portfolio optimization modeling 
language on top of a generalized algebraic modeling language. By focussing on 
a number of different modeling and optimization approaches, the Webinar is 
going to provide new insights for a broad range of interested parties.
}

